seo_agents/requirements.txt
# Ayonne SEO Agent Dependencies
# Minimal dependencies, no paid APIs required

# HTTP requests
requests>=2.31.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.1.0

# YAML configuration
pyyaml>=6.0.1

# JSON schema validation
jsonschema>=4.21.0

# URL parsing and validation
validators>=0.22.0

# Text analysis
textstat>=0.7.3  # Readability scores

# Rate limiting
ratelimit>=2.2.1

# CLI arguments
click>=8.1.0

# Date handling
python-dateutil>=2.8.2

# Logging
colorlog>=6.8.0

# Optional: Google APIs (for GSC integration if available)
# google-api-python-client>=2.118.0
# google-auth>=2.27.0
